refactor token_capacity to size_t
Shouldn't ever be negative anyways.

Fixes the warning "argument 2 range [18446744071562067968, 18446744073709551615] exceeds maximum object size 9223372036854775807 [-Werror=alloc-size-larger-than=]"

+++ lcmgen/tokenize.h
@@ -18,7 +18,7 @@ struct tokenize {
     char *token;
 
     // bytes allocated for token.
-    int token_capacity;
+    size_t token_capacity;
 
     int token_line, token_column;
